Make config-update more robust
Now the files which are not part of the config DB are also saved, so
they can be restored later in case of errors.

It also changes the behaviour when no parameters are
passed. Previously we tried to find a config file in the DB based on
the current rev of the kernel and the name of the config. Now, we just
update ".config" in the kernel repository, so this is basically
equivalent to run "make oldconfig" except that a template can be used.

Signed-off-by: Franck Bui-Huu <fbuihuu@gmail.com>