Track all kernel components inside a "super project"
Now we use a super project to track all components. Making this change
allows to get rid of the release file completely since releases are
described and retrieved by a specific tag/commit inside the super
project.

All kdist-repository management commands are now gone and are replaced
by kdist-project commands which are wrappers around git-submodule.

Only the kernel repository is using a release process, the others
components can have a different tag scheme, we simply use a "snapshot"
of them.

Signed-off-by: Franck Bui-Huu <fbuihuu@gmail.com>