release-project: create a temporary commit so we can find the correct parent
If we're currently on a release (tagged commit) and we're
about to create a new one, then release__get_parent() was not
finding the right parent since parent is the parent of the new
release (not the current one).

So we create a temporary commit to fix this.

It also allows to include the current staged changes in the
log.

Signed-off-by: Franck Bui-Huu <fbuihuu@gmail.com>