whatchanged: rework interface
The interface is now similar to the one that git-cherry has but it
adds the notion of "upstream" and "Mainlined" patch state.

This work also removes any release notions that make the script
lowlevel material. Any release logics are let to the called. Therefore
if you don't specify where the upstream ends, there won't be any
"Mainlined" patches.

kdist-log has been fixed to reflect this change. In its case, it's
pretty simple to find out the bases and kdist-whatchanged will ignore
them if they're not usefull (ie ancestor of merge-base(new, old)).

Signed-off-by: Franck Bui-Huu <fbuihuu@gmail.com>