config__ls_files(): force any tracked config files to have an associated info file
I'm not sure about the previous comment since config info are always
present for any tracked config files.

Morevover 'import' command doesn't use this path except when the
config file already exists (error case) and that doesn't hurt to check
the presence of the info file in that case.

Therefore this patch forces the function to always read the info file.

Signed-off-by: Franck Bui-Huu <fbuihuu@gmail.com>