Config name is now stored in LOCALVERSION field
The main point of this change is that no more kernel source changes
are required to introduce the CONFIG_NAME value in the .config hence
to make kdist work.

The downside is that it's more tricky to figure out what's the config
name is because the user might already use it. Maybe we can force
LOCALVERSION to only contain the config name...

Signed-off-by: Franck Bui-Huu <fbuihuu@gmail.com>