Add bug management support
Bugs can be imported into a a simple database made of text files whose
filenames corresponds to the associated bug ID.

Currenty it understands the CVE official text list provided by the
cve.org that can be found at
http://cve.mitre.org/data/downloads/index.html.

It also uses git notes feature to mark a commit as a fix for a bug
which allows to figure out quickly if a commit is a fix or not. This
is considerably faster when finding out which bugs are fixed by a
specific release (see 'kdist bug log' command).