mdv: use a new scheme for rpm kernel package post installation
Simplify the rpm scriptlet for kernel package: we don't rely in
bootloader-utils (and drakx) to generate a new initrd and to update
bootloader menu entry.

Instead a simple scheme is used: tools that want to be notified when
installing a new kernel should put an executable script in
/etc/kernel/*.d directories. Those scripts will be executed whenever a
kernel binary package is installed or removed.

Also this part was previously x86 specific, but it's now part of the
common arch stuff because tools used by scriptlets are now optional.