Make bug_db__query_by_id() safer and probably faster too (semantic change)
Blindly sourcing bug metadata file can affect kdist local variable values
which can be difficult too debug. The function now only changes the known
bug field values, no more temporary variables are used.

Also we don't call any subshell, so it should be faster.

The change of the metadata file format was required: since we don't
source it anymore we needed to find an appropriate way to separate
field. In order to achieve this, now any fields (with single line
value or not) have a separator which is an empty line, double quotes
are no more used.

Signed-off-by: Franck Bui-Huu <fbuihuu@gmail.com>