Fix rpm-package-<flavour> including kernel source: all devel files must be included
Before this patch, we removed all unwanted architectures. But some devel files
needed during some kbuild operations live in a specific arch directory.
arch/um/scripts/Makefile.rules is an example which is needed by hostfs
when doing a 'make clean'. This is probably a bug but we have to deal
with it.